backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.12
python-dotenv==1.0.1
pytest==8.2.1
pytest-django==4.8.0
black==24.4.2
isort==5.13.2
ruff==0.4.5
openai==1.52.0
pdfplumber==0.11.2
